Refactor local installation methods into their own service
