+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Interfaces/Helpers/ILocalInstallationHelper.cs
@@ -1,0 +1,10 @@
+﻿using CSETWebCore.DataLayer.Model;
+
+namespace CSETWebCore.Helpers
+{
+    public interface ILocalInstallationHelper
+    {
+        void determineIfUpgradedNeededAndDoSo(int newuserID, CSETContext tmpContext);
+        bool IsLocalInstallation();
+    }
+}
