+++ CSETWebApi/CSETWeb_Api/CSETWebCore.Interfaces/Helpers/IUserAuthentication.cs
@@ -8,7 +8,5 @@ namespace CSETWebCore.Interfaces.Helpers
     {
         LoginResponse Authenticate(Login login);
         LoginResponse AuthenticateStandalone(Login login, ITokenManager tokenManager);
-        void determineIfUpgradedNeededAndDoSo(int newuserID, CSETContext tmpContext);
-        bool IsLocalInstallation(String app_code);
     }
 }
